#11 MessageBox UI / 팔로우 수정 / Styles
follower count -> ViewModel로

+++ app/src/main/java/com/dbclass/hanstagram/data/repository/PostRepository.kt
@@ -34,7 +34,7 @@ object PostRepository {
         }
     }
 
-    suspend fun getPostsCount(userID: String): Int {
+    suspend fun getPostsCount(userID: String): Long {
         return CoroutineScope(Dispatchers.Default).async {
             postsDao?.getPostsCount(userID)
         }.await() ?: 0
